refactor: monorepo atoms for internal use
