+++ apps/web/modules/users/views/users-type-public-view.tsx
@@ -2,7 +2,7 @@
 
 import { useSearchParams } from "next/navigation";
 
-import { Booker } from "@calcom/atoms/web";
+import { Booker } from "@calcom/atoms/monorepo";
 import { getBookerWrapperClasses } from "@calcom/features/bookings/Booker/utils/getBookerWrapperClasses";
 import { BookerSeo } from "@calcom/features/bookings/components/BookerSeo";
