+++ packages/features/bookings/components/event-meta/Members.tsx
@@ -1,4 +1,4 @@
-import { useIsPlatform } from "@calcom/atoms/web";
+import { useIsPlatform } from "@calcom/atoms/monorepo";
 import { SchedulingType } from "@calcom/prisma/enums";
 import { UserAvatarGroup, UserAvatarGroupWithOrg } from "@calcom/ui";
