+++ packages/lib/hooks/useLocale.ts
@@ -1,6 +1,6 @@
 import { useTranslation } from "next-i18next";
 
-import { useAtomsContext } from "@calcom/atoms/web";
+import { useAtomsContext } from "@calcom/atoms/monorepo";
 
 export const useLocale = (namespace: Parameters<typeof useTranslation>[0] = "common") => {
   const context = useAtomsContext();
