+++ packages/ui/components/tooltip/Tooltip.tsx
@@ -1,7 +1,7 @@
 import * as TooltipPrimitive from "@radix-ui/react-tooltip";
 import React from "react";
 
-import { useIsPlatform } from "@calcom/atoms/web";
+import { useIsPlatform } from "@calcom/atoms/monorepo";
 import classNames from "@calcom/lib/classNames";
 
 export function Tooltip({
